refactor: Segment 결과 저장 로직 분리
