+++ backend/src/main/java/com/overlang/domain/job/service/JobResultService.java
@@ -1,8 +1,10 @@
 package com.overlang.domain.job.service;
 
+import com.overlang.api.dto.job.JobCallbackRequest;
 import com.overlang.api.dto.ocr.OcrItemResponse;
 import com.overlang.api.dto.segment.SegmentResponse;
 import com.overlang.api.dto.segment.SegmentWordResponse;
+import com.overlang.domain.job.entity.Job;
 import com.overlang.domain.job.repository.JobRepository;
 import com.overlang.domain.learning.repository.LearningContentRepository;
 import com.overlang.domain.ocr.repository.OcrItemRepository;
@@ -12,6 +14,7 @@ import com.overlang.domain.segment.entity.SegmentWord;
 import com.overlang.domain.segment.entity.SegmentWordType;
 import com.overlang.domain.segment.repository.SegmentRepository;
 import com.overlang.domain.segment.repository.SegmentWordRepository;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -94,5 +97,66 @@ public class JobResultService {
     segmentRepository.deleteByJobId(jobId);
     ocrItemRepository.deleteByJobId(jobId);
     learningContentRepository.deleteByJobId(jobId);
+  }
+
+  private static final String WORD_SPLIT_REGEX = "\\s+";
+
+  public void saveSegments(Job job, JobCallbackRequest request) {
+    if (request.segments() == null) return;
+
+    List<Segment> segments =
+        request.segments().stream()
+            .map(
+                s ->
+                    new Segment(
+                        job,
+                        s.startTime(),
+                        s.endTime(),
+                        s.seq(),
+                        s.text(),
+                        s.translatedText(),
+                        s.languageCode()))
+            .toList();
+
+    List<Segment> savedSegments = segmentRepository.saveAll(segments);
+
+    saveSegmentWords(savedSegments);
+  }
+
+  private void saveSegmentWords(List<Segment> segments) {
+    List<SegmentWord> segmentWords =
+        segments.stream().flatMap(segment -> createWordsFromSegment(segment).stream()).toList();
+
+    segmentWordRepository.saveAll(segmentWords);
+  }
+
+  private List<SegmentWord> createWordsFromSegment(Segment segment) {
+    List<SegmentWord> words = new ArrayList<>();
+
+    words.addAll(createWordsFromText(segment, segment.getText(), SegmentWordType.ORIGINAL));
+
+    words.addAll(
+        createWordsFromText(segment, segment.getTranslatedText(), SegmentWordType.TRANSLATION));
+
+    return words;
+  }
+
+  private List<SegmentWord> createWordsFromText(
+      Segment segment, String text, SegmentWordType wordType) {
+
+    if (text == null || text.isBlank()) {
+      return List.of();
+    }
+
+    String[] words = text.split(WORD_SPLIT_REGEX);
+    List<SegmentWord> segmentWords = new ArrayList<>();
+
+    for (int i = 0; i < words.length; i++) {
+      segmentWords.add(
+          new SegmentWord(
+              segment, i + 1, segment.getStartTime(), segment.getEndTime(), words[i], wordType));
+    }
+
+    return segmentWords;
   }
 }
