+++ backend/src/main/java/com/overlang/domain/job/service/JobService.java
@@ -21,12 +21,8 @@ import com.overlang.domain.project.entity.ProjectStatus;
 import com.overlang.domain.project.entity.SourceType;
 import com.overlang.domain.project.repository.ProjectRepository;
 import com.overlang.domain.savedword.repository.SavedWordRepository;
-import com.overlang.domain.segment.entity.Segment;
-import com.overlang.domain.segment.entity.SegmentWord;
-import com.overlang.domain.segment.entity.SegmentWordType;
 import com.overlang.domain.segment.repository.SegmentRepository;
 import com.overlang.domain.segment.repository.SegmentWordRepository;
-import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -214,7 +210,7 @@ public class JobService {
         request.progress(), request.currentStage(), request.errorCode(), request.errorMessage());
 
     jobResultService.deletePreviousResults(job.getId());
-    saveSegments(job, request);
+    jobResultService.saveSegments(job, request);
     saveOcrItems(job, request);
     saveLearningContents(job, request.learningData());
 
@@ -226,28 +222,6 @@ public class JobService {
         request.progress(), request.currentStage(), request.errorCode(), request.errorMessage());
 
     job.getProject().markFailed();
-  }
-
-  private void saveSegments(Job job, JobCallbackRequest request) {
-    if (request.segments() == null) return;
-
-    List<Segment> segments =
-        request.segments().stream()
-            .map(
-                s ->
-                    new Segment(
-                        job,
-                        s.startTime(),
-                        s.endTime(),
-                        s.seq(),
-                        s.text(),
-                        s.translatedText(),
-                        s.languageCode()))
-            .toList();
-
-    List<Segment> savedSegments = segmentRepository.saveAll(segments);
-
-    saveSegmentWords(savedSegments);
   }
 
   private void saveOcrItems(Job job, JobCallbackRequest request) {
@@ -280,43 +254,6 @@ public class JobService {
         blurRegion != null ? blurRegion.y() : null,
         blurRegion != null ? blurRegion.w() : null,
         blurRegion != null ? blurRegion.h() : null);
-  }
-
-  private void saveSegmentWords(List<Segment> segments) {
-    List<SegmentWord> segmentWords =
-        segments.stream().flatMap(segment -> createWordsFromSegment(segment).stream()).toList();
-
-    segmentWordRepository.saveAll(segmentWords);
-  }
-
-  private List<SegmentWord> createWordsFromSegment(Segment segment) {
-    List<SegmentWord> words = new ArrayList<>();
-
-    words.addAll(createWordsFromText(segment, segment.getText(), SegmentWordType.ORIGINAL));
-
-    words.addAll(
-        createWordsFromText(segment, segment.getTranslatedText(), SegmentWordType.TRANSLATION));
-
-    return words;
-  }
-
-  private List<SegmentWord> createWordsFromText(
-      Segment segment, String text, SegmentWordType wordType) {
-
-    if (text == null || text.isBlank()) {
-      return List.of();
-    }
-
-    String[] words = text.split(WORD_SPLIT_REGEX);
-    List<SegmentWord> segmentWords = new ArrayList<>();
-
-    for (int i = 0; i < words.length; i++) {
-      segmentWords.add(
-          new SegmentWord(
-              segment, i + 1, segment.getStartTime(), segment.getEndTime(), words[i], wordType));
-    }
-
-    return segmentWords;
   }
 
   private void saveLearningContents(Job job, CallbackLearningDataRequest learningData) {
